backend/myspinwheelservice/pyproject.toml
[project]
name = "my-spin-wheel-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kirtimanmishra", email = "mishrakirtiman24@gmail.com"},
]
dependencies = [
    "django>=5.1.1",
    "pdm-django>=1.0.0",
    "python-dotenv==1.0.0",
    "djangorestframework==3.15.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
runserver = "python manage.py runserver"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
createsuperuser = "python manage.py createsuperuser"


[tool.pdm]
distribution = true
readme = "README.md"
